version 12-26 材料商材料展示购买页面完成，兼容64-bit

+++ DecorationDesign/Tab1/BookingViewController.m
@@ -97,7 +97,7 @@
 
 -(UITableViewCell*)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    int row = indexPath.row;
+    NSInteger row = indexPath.row;
     if (row<7) {
         static NSString *identifierCell = @"cell0";
         UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:identifierCell];
