+++ DecorationDesign/Tab1/HomeViewController.m
@@ -111,7 +111,7 @@
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    int row=[indexPath row];
+    NSInteger row=[indexPath row];
     
     if (row==0) {
         static NSString *CellIdentifier = @"Cell0";
@@ -305,9 +305,9 @@
     return nil;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    int row=[indexPath row];
+    NSInteger row=[indexPath row];
     switch (row) {
         case 0:
             return 80*widthRate;
@@ -327,7 +327,7 @@
     return 0;
 }
 
-- (void)loadScrollViewWithPage:(int)page2 {
+- (void)loadScrollViewWithPage:(NSInteger)page2 {
     if (page2 < 0) return;
     if (page2 >= kNumberOfPages) return;
     
@@ -425,7 +425,7 @@
 }
 
 - (void)changePage:(id)sender {
-    int page2 = pageControl.currentPage;
+    NSInteger page2 = pageControl.currentPage;
     // load the visible page and the page on either side of it (to avoid flashes when the user starts scrolling)
     [self loadScrollViewWithPage:page2 - 1];
     [self loadScrollViewWithPage:page2];
