+++ DecorationDesign/Tab1/ZuopingLookViewController.m
@@ -156,7 +156,7 @@
     
 }
 
-- (void)loadScrollViewWithPage:(int)page2 {
+- (void)loadScrollViewWithPage:(NSInteger)page2 {
     if (page2 < 0) return;
     if (page2 >= kNumberOfPages) return;
     
@@ -195,7 +195,7 @@
 }
 
 - (void)changePage:(id)sender {
-    int page2 = pageControl.currentPage;
+    NSInteger page2 = pageControl.currentPage;
     // load the visible page and the page on either side of it (to avoid flashes when the user starts scrolling)
     [self loadScrollViewWithPage:page2 - 1];
     [self loadScrollViewWithPage:page2];
